msg_reader/requirements.txt
oci
mysql-connector-python
opentelemetry-api
opentelemetry-sdk